fix repost bug and sorting bug

+++ index.js
@@ -7,62 +7,77 @@ const send = require('gmail-send')({
     user: '',
     pass: '',
     to: [''],
-    subject: '',
+    subject: 'New House Found...',
     text: ''});
 
 const dbAdmin = "";
 const adminPassword = "";
 const logIn = dbAdmin + ":" + adminPassword;
-
 const mongoUrl = "mongodb+srv://"+logIn+"@cluster0.z4lwg.mongodb.net/dbOne?retryWrites=true&w=majority";  //MongoDB Atlas account connection
 const testUrl1 = "https://sfbay.craigslist.org/search/scz/apa?postal=95060&max_price=4300&min_bedrooms=4&availabilityMode=0&sale_date=all+dates";
 const testUrl2 = "https://sfbay.craigslist.org/search/scz/apa?postal=95060&max_price=6000&min_bedrooms=5&availabilityMode=0&sale_date=all+dates";
-const housingPageURL = "https://sfbay.craigslist.org/search/scz/apa?postal=95060&max_price=2000&min_bedrooms=1&availabilityMode=0&sale_date=all+dates";                                        //Craigslist housing page to be scraped
+const housingPageURL = "https://sfbay.craigslist.org/search/scz/apa?postal=95060&availabilityMode=0&sale_date=all+dates"; //Craigslist housing page to be scraped
+const random = (Math.random() + 0.4) * 1000;                              //0.4 - 1.399... seconds
 
 
-async function main() {                                                 //MAIN FUNCTION OF THE PROGRAM
+async function main() {                                                   //MAIN FUNCTION OF THE PROGRAM
+  await connectToMongoDb();
+  const browser = await puppeteer.launch({ headless: true });
   while (true) {
     try {
-      await connectToMongoDb();
-      const browser = await puppeteer.launch({ headless: true });
+      await sleep(random);
       const page = await browser.newPage();
-      let prevListings = await getCollection(Listing);                  //drop the current "listings" collection and return its contents
+      let prevListings = await getCollection(Listing);                    //drop the current "listings" collection and return its contents
       prevListings = await removeKey(prevListings, "_id");
-      let listings = await scrapeListings(page);                        //scrape the "housing home page" listings
-      let newListings = await difference(listings, prevListings);       //find new house listings that don't exist in the "prevListings"...
-      if(newListings.length == 0 && prevListings.length != 0) {         //...but exist in "listings"
+      let listings = await scrapeListings(page);                          //scrape the "housing home page" listings
+      let newListings = await difference(listings, prevListings);         //find new house listings that don't exist in the "prevListings"...
+      if(newListings.length == 0 && prevListings.length != 0) {           //...but exist in "listings"
         console.log("no new listings");
-      }else if(newListings.length != 0 && prevListings.length != 0){    //if there are new listings and this is not the first iteration...
-        console.log("NEW listings:");                                   //...of the program
-        newListings = await scrapeListingsDescriptions(listings, page); //scrape the descriptions of each of the house listing
-        newLlistings = await sortListings(newListings, "daysAgo");      //sort listings by "daysAgo" posted (newest first)
-        console.log("NEW LISTINGS", newListings);
-        await sendListings(newListings);                                //send the the messages via text/email
-        console.log("ALL listings:");
-        listings = await scrapeListingsDescriptions(listings, page);
-        listings = await sortListings(listings, "daysAgo");             //sort listings by "daysAgo" posted (newest first)
-        await Listing.collection.drop();                                //drop the current db collection
-        await saveListings(listings);                                   //save the updated listings into the db collection
+      }else if(newListings.length != 0 && prevListings.length != 0){      //if there are new listings and this is not the first iteration...
+        console.log("NEW listings:");                                     //...of the program
+        newListings = await scrapeListingsDescriptions(newListings, page);//scrape the descriptions of each of the new listing
+        newListings = newListings.filter(newListing => {                  //filter out reposts
+          newListing.daysAgo == 0;
+        });
+        if(newListings.length == 0) {                                   //if all newListings were reposts
+          console.log("no new listings");
+          continue;
+        }else{
+          console.log("NEW LISTINGS", newListings);
+          await sendListings(newListings);                              //send the the new listings via text/email
+          console.log("ALL listings:");
+          listings = await scrapeListingsDescriptions(listings, page);
+          listings = await sortListings(listings);                      //sort listings by "daysAgo" posted (newest first)
+          console.log("listings", listings);
+          await Listing.collection.drop();                              //drop the current db collection
+          await saveListings(listings);                                 //save the updated listings into the db collection
+        }
       }else{
         console.log("ALL listings:");                                   //"first iteration" case of the program
         listings = await scrapeListingsDescriptions(listings, page);
-        listings = await sortListings(listings, "daysAgo");
+        listings = await sortListings(listings);
+        console.log("listings", listings);
         await saveListings(listings);
       }
       await page.close();
-      await browser.close();
-      await mongoose.connection.close();
       console.log("\n");
     } catch(err) {                                  //if main() function fails
       const {result,full} = await send({            //send error message
-        subject: "Program Stopped",
+        subject: "Program Stopped (Error)",
         text: "" + err
       });
       console.log(result);
       console.log(err);
+      await page.close();
+      await browser.close();
+      await mongoose.connection.close();
       process.exit(0);
     }
   }
+}
+
+async function sleep(miliseconds) {
+  return new Promise(resolve => setTimeout(resolve, miliseconds));
 }
 
 async function connectToMongoDb() {                 //FUNCTION TO CONNECT TO MONGODB
@@ -83,8 +98,8 @@ async function getCollection(model) {               //FUNCTION TO RETURN ALL ITE
   try {
     collection = await Listing.find({});            //find all items of a collection
   } catch(err) {
-    collection = [];                                //if error, collection is empty
     console.log(err);
+    collection = [];                                //if error, collection is empty
   } finally {
     return collection;
   }
@@ -103,11 +118,12 @@ async function removeKey(listings, key){            //FUNCTION THAT REMOVES A KE
 
 async function scrapeListings(page) {                             //FUNCTION TO SCRAPE LISTINGS FROM THE CRAIGSLIST HOUSING, MAIN PAGE...
   console.log("scraping listings page...");                       //...AND RETURN THE LISTINGS AS AN ARRAY OF OBJECTS
+  let listings = [];
   try {
     await page.goto(housingPageURL);                              //access a craiglist housing page via url
     const html = await page.content();
     const $ = cheerio.load(html);
-    const listings = $(".result-info").map((index, element) => {  //loop through all listings via "result-info" element and store...
+    listings = $(".result-info").map((index, element) => {        //loop through all listings via "result-info" element and store...
       const timeElement = $(element).find(".result-date");        //...each return value in an array
       const titleElement = $(element).find(".result-title");      //query the html page for all the desired elements
       const resultPriceElement = $(element).find(".result-meta").find(".result-price");
@@ -133,9 +149,12 @@ async function scrapeListings(page) {                             //FUNCTION TO 
         url
       };
     }).get();
-    return listings;                              //return the array of scraped listings
   } catch(err) {
     console.log(err);                             //log error if scraping listings fails
+    console.log("err entered");
+    listings = [];
+  } finally {
+    return listings;                              //return the array of scraped listings
   }
 }
 
@@ -161,7 +180,6 @@ async function scrapeListingsDescriptions(listings, page) {    //FUNCTION TO SCR
       listings[i].daysAgo = daysAgo(listings[i], daysAgoElement);
       listings[i].address = $("div.mapaddress").text();
       listings[i].specs = $format($, specsElements);
-      const random = (Math.random() + 0.4) * 1000;             //0.4 - 1.399... seconds
       await sleep(random);                                     //buffer in order to prevent being blocked from craigslist
     }
     return listings;
@@ -202,15 +220,11 @@ function $format($, elements) {                      //FUNCTION TO PARSE THE TEX
   }
 }
 
-async function sleep(miliseconds) {
-  return new Promise(resolve => setTimeout(resolve, miliseconds));
-}
-
-function sortListings(listings, key) {              //FUNCTION TO SORT LISTINGS
+function sortListings(listings) {                   //FUNCTION TO SORT LISTINGS by "daysAgo" key value
   console.log("sorting...");
   try {
     listings.sort((a, b) => {                       //sort listings by "daysAgo" key value
-      return a.key - b.key;
+      return a.daysAgo - b.daysAgo;
     });
     return listings;
   } catch(err) {
